metodos con hibernate agregados

+++ src/main/java/com/grupo4/integrador/dto/TurnoDto/CrearTurnoDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class CrearTurnoDto {
-    private int odontologoId;
-    private int pacienteId;
+    private Long odontologoId;
+    private Long pacienteId;
     private String fecha;
 }
